requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
SQLAlchemy==2.0.29
pydantic==2.7.1
pydantic-settings==2.2.1
passlib[bcrypt]==1.7.4
python-jose==3.3.0
alembic==1.13.2
python-multipart==0.0.9
httpx==0.27.0
pytest==8.2.0
